Creada una clase dedicada, ParserController.kt (codigo desplazado desde ResumenController.kt) Agregada llamada al metodo init de esa clase al Main.kt
