+++ src/main/kotlin/Main.kt
@@ -1,10 +1,12 @@
 import controllers.ResumenController
 import controllers.DirController
 import controllers.DistritoController
+import controllers.ParserController
 
 fun main() {
 
     DirController.init()
+    ParserController.init()
     ResumenController.init()
     DistritoController.init()
